+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Project : Onchange on Project.isInvoicingExpenses and Project.isInvoicingPurchases
 - TeamTask : Added new fields "Private","internalDesription" and "orderAccepted".
 - Axelor-Business-Support : Addition of new module 'axelor-business-support'.
+- TeamTask : Return ticket functionality added.
 
 ## Improvements
 - TimesheetLine: Adding M2O field TeamTask and integer field timeToCharge.